pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent_builder"
version = "0.1.0"
description = "A modular Python framework for building AI agents"
authors = [
    { name = "Your Name", email = "your@email.com" }
]
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "requests",
    "openai",
    "anthropic",
    "python-dotenv"
]

[tool.setuptools.packages.find]
where = ["agent_builder"] 